bear::rolling_credits supports bounded text and alignment.
The is not drawn outside the item. If a line is too long,
a line break is inserted. The multiple lines are still
displayed as one line regarding the fading and the
movement.

The text can be aligned horizontally using the align comments
in the source file:
  #align {align_left,align_center,aling_right}

Default is align_center.