[GL] Includes GLES headers.
A #define is added to replace GLdouble by GLfloat, as the former is not
available in OpenGL ES.

The error cases in gl_error.cpp are enclosed by #define in order to test if
the error codes are available on the platform or not.

Numeric constants are casted to GLdouble in std::{min,max} calls.