[SDL2] Update screen management with SDL 2's API.
SDL_SetVideoMode() does not exist anymore. We use SDL_CreateWindow() instead,
followed by SDL_GL_CreateContext.

We keep the pointer returned by this CreateWindow to get the size of the window
in the places where SDL_GetVideoInfo() and SDL_GetVideoSurface where used.

m_display_size is not needed anymore. We use SDL_GetDesktopDisplayMode() to find
the initial screen resolution.

Finding the most suitable screen resolution is done with SDL_GetDisplayMode().

The event processing uses the new SDL_PeepEvents() interface and searches only
for SDL_QUIT. I am not sure that the old loop was not incorrect.