Finalisation of the flag allowing to reset an animation with an action, or not.
The option is available in the mark properties frame.

Fix how its value was read in the model loader. It must be read before the index
of the animation.

The version of the editors is now 0.10.0.

Fix the swap and copy functions of bear::engine::model_mark. Some members were
not set.