The game is paused when the application goes to the background.
bear::engine::game_local_client can now be paused and woke up.
When it goes to sleep, the level is paused and the music muted.

The game goes to sleep when the application goes to the background.
It is detected via SDL events, handled in the new class
android_system_event_manager, which implements base_system_event_manager.

The game instance creates the event manager during its initialization.