The sounds are loaded in a thread.
Half the time spent loading a level is in the loading of the level's music. This
commit moves the loading of the sounds in a thread in order to allow continuing
the process without blocking.

If one tries to play the sound before the loading completes, then the thread is
forced to finish.