Clean up sleep/wake up state transition.
default_system_event_manager has been added in order to process the events
targetting the window. The game goes to sleep when the window is minimized and
is waken up when the window is restored. This class handles the quit event
too, thus we can remove the old behavior consisting in ending the game if
screen::end_render() returned true.

SDL events are now pumped in the {android,default}_system_event_manager, in
response to a call from game_local_client. It is more clear than the previous
behavior where the events were processed in response to a call somewhere in
bear::visual.