Fix memory problems.
The sound data copied in the copy constructor of sdl_sound is now deleted[]
in the destructor (SDL Mixer does not free the sound data when the chunk has
been created with QuickLoad).

bear::base_ground did not initialize its m_middle_set_contact flag.

The thread in gl_renderer may have continued after the destruction of the
renderer. Now we call join to ensure it is finished before deleting the
renderer.